update validatiop for pipeline

+++ src/services/data_pipelines/ingest/save_file.py
@@ -22,9 +22,11 @@ def _split_good_and_bad(df: pd.DataFrame) -> (pd.DataFrame, pd.DataFrame):
 def run_save_file(validated_result: ValidatedResult) -> None:
     file_name = validated_result.file_path.split("/")[-1]
     file_name = f"{file_name}_{get_unique_id()}.csv"
-    good_df, bad_df = _split_good_and_bad(validated_result.final_df)
-    if not good_df.empty:
-        _save_file(f"{data_path_configs.GOOD_DATA_PATH}/{file_name}", good_df)
-    if not bad_df.empty:
-        _save_file(f"{data_path_configs.BAD_DATA_PATH}/{file_name}", bad_df)
-    DirectoryManager.delete_file(validated_result.file_path)
+    # Only do in case csv is parsed properly
+    if validated_result.final_df:
+        good_df, bad_df = _split_good_and_bad(validated_result.final_df)
+        if not good_df.empty:
+            _save_file(f"{data_path_configs.GOOD_DATA_PATH}/{file_name}", good_df)
+        if not bad_df.empty:
+            _save_file(f"{data_path_configs.BAD_DATA_PATH}/{file_name}", bad_df)
+        DirectoryManager.delete_file(validated_result.file_path)
